fix: moved password update from users resource to auth and changed the endpoint
