+++ api/v1/routes/user.py
@@ -6,9 +6,7 @@ from sqlalchemy.orm import Session
 from api.utils.success_response import success_response
 from api.v1.models.user import User
 from api.v1.schemas.user import (
-    DeactivateUserSchema,
-    ChangePasswordSchema,
-    ChangePwdRet, AllUsersResponse, UserUpdate,
+    AllUsersResponse, UserUpdate,
     AdminCreateUserResponse, AdminCreateUser
 )
 from api.db.database import get_db
@@ -51,20 +49,6 @@ async def delete_account(request: Request, db: Session = Depends(get_db), curren
         status_code=200,
         message='User deleted successfully',
     )
-
-
-@user_router.patch("/me/password", status_code=200)
-async def change_password(
-    schema: ChangePasswordSchema,
-    db: Session = Depends(get_db),
-    user: User = Depends(user_service.get_current_user),
-):
-    """Endpoint to change the user's password"""
-
-    user_service.change_password(schema.old_password, schema.new_password, user, db)
-
-    return success_response(status_code=200, message="Password changed successfully")
-
 
 @user_router.patch("/",status_code=status.HTTP_200_OK)
 def update_current_user(
